fix(geomoments): center 2D central moments on the true fractional centroid
centralMom() computed the origin as (int)originOfX / (int)originOfY,
truncating the fractional centroid to whole pixels. That shifted the
origin and corrupted every central, normalized, and Hu moment (shape and
intensity), most visibly making mu10/mu01 non-zero when they are 0 by
definition. The _nt sibling already used the untruncated centroid; this
brings both overloads in line.

Testing:
- Regenerate the pinned central/normalized/Hu goldens against an
  independent numpy/skimage oracle (new
  tests/vetting/oracles/gen_geomoments_analytic.py, token: analytic);
  raw spatial moments are unaffected and unchanged.
- Add an asymmetric L-shape fixture whose centroid is a non-half
  fraction, so the truncation is caught at every moment order rather
  than only the symmetric-rectangle cases.
- Add an analytic invariant test asserting mu10 = mu01 = 0 for both
  shape and intensity moments.

+++ src/nyx/features/2d_geomoments_basic.cpp
@@ -185,7 +185,7 @@ double BasicGeomoms2D::centralMom(const pixcloud& cloud, int p, int q)
     for (auto& pxl : cloud)
     {
         double I = INTEN(double(pxl.inten));
-        sum += I * int_pow(double(pxl.x - baseX) - (int)originOfX, p) * int_pow(double(pxl.y - baseY) - (int)originOfY, q);
+        sum += I * int_pow(double(pxl.x - baseX) - originOfX, p) * int_pow(double(pxl.y - baseY) - originOfY, q);	// fix: center on the true fractional centroid; the (int) truncation shifted the origin and corrupted central/normalized/Hu moments (the _nt sibling already does this)
     }
     return sum;
 }
@@ -198,7 +198,7 @@ double BasicGeomoms2D::centralMom(const pixcloud& cloud, const reintenvec& reali
     for (size_t i = 0; i < n; i++)
     {
         auto& pxl = cloud[i];
-        sum += realintens[i] * int_pow(double(pxl.x - baseX) - (int)originOfX, p) * int_pow(double(pxl.y - baseY) - (int)originOfY, q);
+        sum += realintens[i] * int_pow(double(pxl.x - baseX) - originOfX, p) * int_pow(double(pxl.y - baseY) - originOfY, q);	// fix: center on the true fractional centroid; the (int) truncation shifted the origin and corrupted central/normalized/Hu moments (the _nt sibling already does this)
     }
     return sum;
 }
